Use distinct deployment names.
