+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/ejb/remote/GlobalAuthContextRemoteStatelessEJBFailoverTestCase.java
@@ -7,6 +7,9 @@ package org.jboss.as.test.clustering.cluster.ejb.remote;
 
 import java.util.function.UnaryOperator;
 
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.container.test.api.TargetsContainer;
+import org.jboss.shrinkwrap.api.Archive;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.wildfly.security.auth.client.AuthenticationContext;
@@ -17,6 +20,20 @@ import org.wildfly.security.auth.client.AuthenticationContext;
  * @author Paul Ferraro
  */
 public class GlobalAuthContextRemoteStatelessEJBFailoverTestCase extends AuthContextRemoteStatelessEJBFailoverTestCase {
+    private static final String MODULE_NAME = GlobalAuthContextRemoteStatelessEJBFailoverTestCase.class.getSimpleName();
+
+    @Deployment(name = DEPLOYMENT_1, managed = false, testable = false)
+    @TargetsContainer(NODE_1)
+    public static Archive<?> createDeploymentForContainer1() {
+        return createDeployment(MODULE_NAME);
+    }
+
+    @Deployment(name = DEPLOYMENT_2, managed = false, testable = false)
+    @TargetsContainer(NODE_2)
+    public static Archive<?> createDeploymentForContainer2() {
+        return createDeployment(MODULE_NAME);
+    }
+
     private static AuthenticationContext previousContext;
 
     @BeforeAll
@@ -31,6 +48,6 @@ public class GlobalAuthContextRemoteStatelessEJBFailoverTestCase extends AuthCon
     }
 
     public GlobalAuthContextRemoteStatelessEJBFailoverTestCase() {
-        super(UnaryOperator.identity());
+        super(MODULE_NAME, UnaryOperator.identity());
     }
 }
